OBPIH-7361 Fix single product dialog in transaction report  (#5358)
