+++ grails-app/services/org/pih/warehouse/inventory/InventorySnapshotService.groovy
@@ -425,251 +425,4 @@ class InventorySnapshotService {
         )
         log.info "Updated ${results} inventory snapshots for product ${product}"
     }
-
-    def getTransactionReportData(Location location, Date startDate, Date endDate) {
-
-        def transactionCodes = [TransactionCode.DEBIT, TransactionCode.CREDIT].collect { it.toString() }
-        // Get all transactions between start and end dates for the given location
-        def transactions = TransactionFact.createCriteria().list {
-            projections {
-                productKey {
-                    groupProperty("productCode")
-                }
-                transactionTypeKey {
-                    groupProperty("transactionCode")
-                    groupProperty("transactionTypeName")
-                }
-                sum("quantity")
-            }
-            transactionTypeKey {
-                'in'("transactionCode", transactionCodes)
-            }
-            transactionDateKey {
-                between("date", startDate, endDate)
-            }
-            locationKey {
-                eq("locationId", location.id)
-            }
-        }
-
-        // Transform transaction facts
-        transactions = transactions.collect {
-            [
-                    productCode        : it[0],
-                    transactionCode    : it[1],
-                    transactionTypeName: it[2],
-                    quantity           : it[3]
-            ]
-        }
-        return transactions
-    }
-
-    Integer calculateBalance(List<TransactionEntry> transactionEntries, Integer balance) {
-        List<TransactionEntry> credits = getCreditTransactionEntries(transactionEntries)
-        List<TransactionEntry> debits = getDebitTransactionEntries(transactionEntries)
-        Integer quantityFromCredits = credits.sum { Math.abs(it.quantity) } as Integer ?: 0
-        Integer quantityFromDebits = debits.sum { Math.abs(it.quantity) } as Integer ?: 0
-
-        return balance - quantityFromCredits + quantityFromDebits
-    }
-
-    List<TransactionEntry> getCreditTransactionEntries(List<TransactionEntry> transactionEntries) {
-        return transactionEntries.findAll {
-            it.transaction.transactionType.transactionCode == TransactionCode.CREDIT && it.quantity > 0
-        }
-    }
-
-    List<TransactionEntry> getDebitTransactionEntries(List<TransactionEntry> transactionEntries) {
-        return transactionEntries.findAll {
-            it.transaction.transactionType.transactionCode == TransactionCode.DEBIT ||
-                    (it.transaction.transactionType.transactionCode == TransactionCode.CREDIT && it.quantity < 0)
-        }
-    }
-
-    List<TransactionEntry> getFilteredTransactionEntries(
-            List<TransactionCode> transactionCodes,
-            Date startDate,
-            Date endDate,
-            List<Category> categories,
-            List<Tag> tagsList,
-            List<ProductCatalog> catalogsList,
-            Location location,
-            String orderBy
-    ) {
-        return TransactionEntry.createCriteria().list {
-            inventoryItem {
-                product {
-                    if (categories) {
-                        'in'('category', categories)
-                    }
-
-                    if (tagsList) {
-                        tags {
-                            'in'("id", tagsList.id)
-                        }
-                    }
-
-                    if (catalogsList) {
-                        productCatalogItems {
-                            productCatalog {
-                                'in'("id", catalogsList*.id)
-                            }
-                        }
-                    }
-                }
-            }
-            transaction {
-                if (transactionCodes) {
-                    transactionType {
-                        'in'('transactionCode', transactionCodes)
-                    }
-                }
-
-                if (startDate) {
-                    gt('transactionDate', startDate)
-                }
-
-                if (endDate) {
-                    lt('transactionDate', endDate)
-                }
-
-                if (location) {
-                    inventory {
-                        eq('warehouse', location)
-                    }
-                }
-
-                if (orderBy) {
-                    order(orderBy, 'desc')
-                }
-            }
-        } as List<TransactionEntry>
-    }
-
-    Map<Product, Map<String, Integer>> getDetailedTransactionReportData(Map<Product, List<TransactionEntry>> transactionEntries) {
-        return transactionEntries.collectEntries { product, entriesForProduct ->
-            Map<String, Integer> totalsByType = entriesForProduct
-                    .groupBy { entry ->
-                        entry.transaction.transactionType.name
-                    }
-                    .collectEntries { transactionTypeName, entriesByType ->
-                        Integer total = entriesByType.sum { entry ->
-                            Math.abs(entry.quantity as Integer)
-                        }
-                        [(transactionTypeName): total]
-                    }
-
-            [(product): totalsByType]
-        }
-    }
-
-    List<Object> getTransactionReport(Location location, List<Category> categories, List<Tag> tagsList, List<ProductCatalog> catalogsList, Date startDate, Date endDate, Boolean includeDetails) {
-        List<TransactionCode> adjustmentTransactionCodes = [
-                TransactionCode.CREDIT,
-                TransactionCode.DEBIT
-        ]
-
-        // Transaction entries that have relation to the transactions happened between startDate <-> endDate
-        // with appropriate filter applied and with transaction type code that is credit or debit
-        List<TransactionEntry> transactionEntriesWithinDateRange = getFilteredTransactionEntries(
-                adjustmentTransactionCodes,
-                startDate,
-                endDate,
-                categories,
-                tagsList,
-                catalogsList,
-                location,
-                null
-        )
-        // Transaction entries that have relation to the transactions happened between endDate <-> today
-        // with appropriate filter and sorting by transaction date applied
-        List<TransactionEntry> transactionsEntriesAfterEndDate = getFilteredTransactionEntries(
-                adjustmentTransactionCodes,
-                endDate,
-                null,
-                categories,
-                tagsList,
-                catalogsList,
-                location,
-                'transactionDate'
-        )
-
-        // Grouping transaction entries by product to get the desired report granularity
-        Map<Product, List<TransactionEntry>> productsMap = transactionEntriesWithinDateRange.groupBy {
-            it.inventoryItem.product
-        }
-        Map<Product, List<TransactionEntry>> productsMapAfterEndDate = transactionsEntriesAfterEndDate.groupBy {
-            it.inventoryItem.product
-        }
-
-        // QoH available at the time of running the report - the closing balance is calculated by
-        // adding / subtracting all of the credits / debits that happened between endDate <-> today
-        Map<Product, Integer> initialQuantityForBalanceCalculations = !productsMap.isEmpty()
-                ? productAvailabilityService.getQuantityOnHandByProduct(location, productsMap.keySet().toList())
-                : [:]
-
-        // AvailableTransactionTypes and detailedReportData are only used in case of generating CSV.
-        // The CSV file should contain additional information about the product and the transaction
-        // entries should be grouped by transaction types (greater granularity)
-        List<TransactionType> availableTransactionTypes = includeDetails
-            ? TransactionType.createCriteria().list { 'in'("transactionCode", adjustmentTransactionCodes) }
-            : []
-
-        Map<Product, List<Integer>> detailedReportData = includeDetails
-            ? getDetailedTransactionReportData(productsMap)
-            : [:]
-
-        // Final calculations of data:
-        // 1. Get the current QoH
-        // 2. Calculate closing balance using transaction entries between endDate <-> today
-        // 3. Calculate opening balance using transaction entries between startDate <-> endDate
-        // Additional calculation info:
-        // 1. CREDITS = transaction entries in relation with transactions that are CREDIT type
-        // and the quantity of that transaction entry is greater than 0
-        // 2. DEBITS = transaction entries in relation with transaction that are DEBIT type
-        // and transaction that are CREDIT type, but with quantity lower than 0
-        return productsMap.collect { key, value ->
-            List<TransactionEntry> entriesAfterEndDate = productsMapAfterEndDate[key] ?: []
-            Integer initialQuantity = initialQuantityForBalanceCalculations[key] ?: 0
-            Integer closingBalance = entriesAfterEndDate.size()
-                    ? calculateBalance(entriesAfterEndDate, initialQuantity)
-                    : initialQuantity
-            Integer openingBalance = calculateBalance(value, closingBalance)
-            Integer credits = getCreditTransactionEntries(value).sum { it.quantity } as Integer ?: 0
-            Integer debits = getDebitTransactionEntries(value).sum { Math.abs(it.quantity) } as Integer ?: 0
-            // In the new version of the report, it's not based on the inventory snapshot.
-            // So we don't have to calculate it in the following way:
-            // closingBalance - openingBalance - credits + debits,
-            // because the data is accurate, so we can just compare
-            // the closing and opening balance
-            Integer adjustments = closingBalance - openingBalance
-
-            return [:].with {
-                it["Code"] = key.productCode
-                it["Name"] = key.name
-                if (includeDetails) {
-                    it["Product Family"] = key.productFamily?.name ?: ''
-                }
-                it["Display Name"] = key?.displayName ?: ''
-                it["Category"] = key.category.name
-                if (includeDetails) {
-                    it["Formulary"] = key.productCatalogsToString()
-                    it["Tag"] = key.tagsToString()
-                }
-                it["Unit Cost"] = key.pricePerUnit ?: ''
-                it["Opening"] = openingBalance
-                it["Credits"] = credits
-                it["Debits"] = debits
-                if (includeDetails) {
-                    availableTransactionTypes.each{ transactionType ->
-                        String columnName = LocalizationUtil.getLocalizedString(transactionType?.name)
-                        it[columnName] = detailedReportData[key]?.get(transactionType?.name) ?: 0
-                    }
-                }
-                it["Adjustments"] = adjustments
-                it["Closing"] = closingBalance
-                it
-            }
-        }
-    }
 }
